requirements.txt
structlog
PyYAML
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72

langchain-groq==0.3.6
langchain-google-genai==2.1.8
sentence-transformers
pypdf 
python-docx 
faiss-cpu
uvicorn==0.35.0
fastapi==0.116.1
PyMuPDF==1.26.3
pypdf==5.8.0
slowapi
python-multipart
-e .